refactor team member component
