--- Web/starter_project_barbies/types/team/TeamMemberCardProps.ts
@@ -1,9 +1,0 @@
-import { SocialMediaLink } from './SocialMediaLink'
-
-export type TeamMemberCardProps = {
-  name: string
-  jobTitle: string
-  description: string
-  image: string
-  links: SocialMediaLink[]
-}
