+++ Web/starter_project_barbies/types/team/team-member-card-props.ts
@@ -1,0 +1,5 @@
+import { TeamMember} from "@/types/team/team-member";
+
+export type TeamMemberCardProps = {
+  teamMember: TeamMember
+}
