+++ Web/starter_project_barbies/types/team/team-member.ts
@@ -1,0 +1,10 @@
+import { SocialMediaLink } from "@/types/team/social-media-link";
+
+export type TeamMember = {
+  id: string,
+  name: string
+  jobTitle: string
+  description: string
+  image: string
+  links: SocialMediaLink[]
+}
